requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Naked==0.1.31
Pillow==7.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.11.2
